// web/src/openapi/model/event.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DOCUMENTATION
 * OpenAPI spec version: 1.0.0
 */
import type { EventCover } from './eventCover';
import type { EventCreatedBy } from './eventCreatedBy';
import type { EventEventType } from './eventEventType';
import type { EventFlyer } from './eventFlyer';
import type { EventUniverse } from './eventUniverse';
import type { EventUpdatedBy } from './eventUpdatedBy';
import type { Programme } from './programme';

export interface Event {
  id?: string | number;
  documentId?: string;
  title: string;
  summary?: string;
  cover: EventCover;
  universe?: EventUniverse;
  description: string;
  startAt: string;
  endAt: string;
  flyer?: EventFlyer;
  slug: string;
  displayTitle?: string;
  location?: string;
  organizers?: string;
  online: boolean;
  facebookLink?: string;
  programmes?: Programme[];
  eventType: EventEventType;
  createdAt?: string;
  updatedAt?: string;
  publishedAt?: string;
  createdBy?: EventCreatedBy;
  updatedBy?: EventUpdatedBy;
  locale?: string;
  localizations?: Event[];
}